admin apis: handle error of params on admin apis

+++ admin/api/deleteDegree.php
@@ -6,6 +6,16 @@ require "../../config/database.php";
 
 header('Content-Type: application/json');
 
+if (!isset($_POST['degreeId'])) {
+    $result = array(
+        'success' => false,
+        'message' => 'Require degreeId'
+    );
+
+    echo json_encode($result);
+    exit;
+}
+
 $degreeId = $_POST['degreeId'];
 
 $sql = "DELETE FROM degree WHERE degreeId = $degreeId";
